fix xslt3 word numbering on 32-bit (#1458)
* fix xslt3 word numbering on 32-bit

* fix xslt3 start-at integer parsing

* fix xsl:number integer formatting

* test: deduplicate ordinal flag literal

* fix big-number ordinal formatting

* fix unicode digit minimum integer formatting

+++ xslt3/number_words_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 func TestFrenchCardinalTens(t *testing.T) {
 	cases := []struct {
-		n    int
+		n    int64
 		want string
 	}{
 		{40, "quarante"},
